+++ api/src/opentrons/protocol_api/__init__.py
@@ -4,7 +4,7 @@ This package defines classes and functions for access through a protocol to
 control the OT2.
 
 """
-from .definitions import MAX_SUPPORTED_VERSION  # noqa(F401)
+from opentrons.protocols.api_support.definitions import MAX_SUPPORTED_VERSION  # noqa(F401)
 from . import labware  # noqa(E402)
 from .contexts import (ProtocolContext,  # noqa(E402)
                        InstrumentContext,
