+++ api/src/opentrons/protocols/api_support/definitions.py
@@ -1,6 +1,6 @@
 import abc
 
-from ..protocols import types
+from opentrons.protocols import types
 
 MAX_SUPPORTED_VERSION = types.APIVersion(2, 7)
 #: The maximum supported protocol API version in this release
