+++ api/tests/opentrons/protocol_api/test_labware.py
@@ -3,15 +3,16 @@ import json
 import pytest
 
 from opentrons.protocol_api import (
-    labware, MAX_SUPPORTED_VERSION, module_geometry)
+    labware, MAX_SUPPORTED_VERSION)
+from opentrons.protocols.api_support import module_geometry
 
 from opentrons_shared_data import load_shared_data
 from opentrons.calibration_storage import (
     helpers, get, delete, file_operators)
 from opentrons.types import Point, Location
 from opentrons.protocols.types import APIVersion
-from opentrons.protocol_api.geometry import Deck
-from opentrons.protocol_api.module_geometry import (
+from opentrons.protocols.api_support.geometry import Deck
+from opentrons.protocols.api_support.module_geometry import (
     ModuleGeometry, MagneticModuleModel, ModuleType)
 
 test_data = {
